requirements.txt
azure-servicebus
flanker
email-validator